Major changes, read:
All of the function handlers were moved to main, since they are
application logic, will have dependencies, and may need to be configured
at runtime without the use of global variables (yuck).

The "JWT" was improved to not use Gorilla and the JWT/Cookie has been
spun off into it's own module, the goal is to encourage an architecture
and example set which is document well in the code herein.